use index dates to fetch updated data files

+++ lib/fetch.rb
@@ -6,8 +6,8 @@ BaseUrl = "http://onehouronelife.com/publicLifeLogData/"
 
 $http = HTTPClient.new
 
-def fetch_file(path, cache_path)
-  if File.exist?(cache_path)
+def fetch_file(path, cache_path, date = Time.at(0))
+  if File.exist?(cache_path) && date < File.mtime(cache_path)
     return File.open(cache_path, "r") {|f| f.read}
   else
     contents = $http.get_content(BaseUrl + path)
@@ -20,14 +20,15 @@ def fetch_file(path, cache_path)
   end
 end
 
-server_directory = fetch_file("", "cache/index.html")
+server_directory = fetch_file("", "cache/index.html", Time.now)
 
 def extract_path_list(directory)
   paths = []
   Nokogiri::HTML(directory).css('a').each do |node|
     path = node.attr(:href)
     next if path == '../' or path == 'lifeLog/'
-    paths << path
+    date = DateTime.parse(node.next.content.strip.chop.strip).to_time
+    paths << [path, date]
   end
 
   return paths
@@ -35,15 +36,14 @@ end
 
 server_list = extract_path_list(server_directory)
 
-server_list.each do |path|
+server_list.each do |path,date|
   FileUtils.mkdir_p('cache/' + path)
 
-  index = fetch_file(path, "cache/#{path}/index.html")
+  index = fetch_file(path, "cache/#{path}/index.html", date)
 
   log_paths = extract_path_list(index)
-  log_paths.each do |log_path|
+  log_paths.each do |log_path,log_date|
     p log_path
-    fetch_file("#{path}#{log_path}", "cache/#{path}#{log_path}")
+    fetch_file("#{path}#{log_path}", "cache/#{path}#{log_path}", log_date)
   end
 end
-
